fix(common): added logic to handle image orientation during thumbnail generation
