+++ packages/common/src/index.js
@@ -7,6 +7,7 @@ export {default as base64} from './base64';
 export {default as cappedDebounce} from './capped-debounce';
 export {default as checkRequired} from './check-required';
 export {default as Defer} from './defer';
+export {default as imageOrientationUtil} from './image-orientation-util';
 export {default as makeStateDataType} from './make-state-datatype';
 export {default as make} from './template-container';
 export {default as oneFlight} from './one-flight';
